regpg: simplify ansible gpg-preload.yml playbook
Do not bother checking for the GPG_AGENT_INFO environment
variable. If I understand the gpg-2.1 documentation, it
starts the agent on demand, so the environment variable is
no longer needed.

The minor downside of this change is the nasty error messages
when the agent isn't available but gpg stupidly tries several
times to use it.