test: compatibility with gpg-2.2 empty public keyrings
Unlike earlier versions, gpg-2.2 doesn't reduce an empty
public keyring to zero bytes. This is OK, except it
confuses our test for the hack that forces gpg-2.1
to use gpg-1 compatible public keyring files.

This fix is in the tests not in regpg itself because
gpg's change in behaviour is benign from regpg's point
of view.