regpg: the ansible-vault --output story is even sadder than I thought
It looks like ansible-vault 2.4.0.0 is trying to canonicalize the output
pathname, and in the case of a pipe it screws it up spectacularly.

The documentation (still) says `--output -` should write to stdout, but
instead it writes to the file `./-`.

If you use `--output /dev/stdout` when the output is a terminal,
it works.

If you write `ansible-vault decrypt --output /dev/stdout secret | cat`
it explodes like:

ERROR! Unexpected Exception, this is probably a bug: [Errno 2] No such file or directory: u'/proc/8528/fd/pipe:[3490448]'