ansible: attempt to work around gpg2 reliability problems
This is related to the Debian bugs linked below. I don't think I have
encountered these problems using regpg interactively (hence no change
to the main code) but when Ansible is hammering gpg concurrently, the
failure rate is enough to be annoying.

This tries gpg1 first, which might fail if it is not installed or
if the secret keys are in the new format. Then it has two tries
with (plain) gpg, in case the first one fails spuriously.

This might do unpleasant things to intended failures. I'll see
what happens in practice and amend if necesary.

https://bugs.debian.org/868550
https://bugs.debian.org/902316