Attempt to make the Ansible filter work on Python 3
Ansible now aims to support Python3 on both the controler and
in modules, and MacOS Homebrew (at least for me) installs Ansible
running under Python3.

The gpg_d filter doesn't then seem to work, I think because
it is returning a byte string rather than a native string. This patch
aims to fix this by reference to code in action.py and
https://docs.ansible.com/ansible/latest/dev_guide/developing_python_3.html

It's been minimally tested, but does seem to fix the problem for Python3
under MacOS. I don't know if it breaks things under Python2.