test/ansible: check both python2 and python3
Also, check that the gpg_d filter works with every version, by
re-running the gpg-preload playbook.

Also, check that the gpg retry code in the plugins works OK when gpg1
is broken.

There is a bug in Debian 9's version of Python 3 that relatively
frequently causes a lot of spurious exception traceback spew on
stderr, so the "stderr quiet" tests have been amended to ignore that
specific exception. (https://github.com/ansible/ansible/issues/21982)

The version tag in the test output includes the Python version, and
the stderr and stdout checks now include the version tag.

The current devel branch of Ansible complains about the system python
being untrustworthy, so we amend ansible.cfg to suppress those warnings.