ReGPG::Login: simplified basic auth support
The trailing options argument to `read_login` was kind of ugly, and
for code that needs options it isn't much harder to use the `new`
constructor. So, get rid of options support in `read_login` (leaving
only the check list) and also remove the `basic` alias for
`auth_basic` to make it more clear what is the purpose of the option.

This is a breaking change, but ipreg is the only user at the moment,
and I have refactored the ipreg code to use `new` where it needed the
`auth_basic` option.